moved more code generation functionality to `common`
