+++ crates/intrinsic-test/src/common/mod.rs
@@ -13,6 +13,7 @@ pub mod indentation;
 pub mod intrinsic;
 pub mod intrinsic_helpers;
 pub mod values;
+pub mod write_file;
 
 /// Architectures must support this trait
 /// to be successfully tested.
